Fix a warning
D:\a\harfbuzz\harfbuzz\src\hb-bit-set-invertible.hh(83): warning C4805: '^': unsafe mix of type 'uint32_t' and type 'const bool' in operation
[870/1075] Compiling C++ object src/harfbuzz.dll.p/hb-face.cc.obj
cl : Command line warning D9025 : overriding '/EHs' with '/EHs-'
cl : Command line warning D9025 : overriding '/EHc' with '/EHc-'

+++ src/hb-bit-set-invertible.hh
@@ -80,7 +80,7 @@ struct hb_bit_set_invertible_t
     next (&v);
     return v == INVALID;
   }
-  uint32_t hash () const { return s.hash () ^ inverted; }
+  uint32_t hash () const { return s.hash () ^ (uint32_t) inverted; }
 
   hb_codepoint_t get_min () const
   {
